二级评论功能接口实现

+++ back/src/main/java/com/su/community/dto/CommentDTO.java
@@ -8,7 +8,7 @@ public class CommentDTO {
     private Long id;
     private Long parentId;
     private Integer type;
-    private Integer commentator;
+    private Long commentator;
     private Long gmtCreate;
     private Long gmtModified;
     private Long likeCount;
